+++ SearchableBuilder.h
@@ -18,7 +18,7 @@ template <typename P>
 class SearchableBuilder {
  private:
     P* buildMatrixState(string state_str);
-    P* buildMatrixCell(int x,int y, double cost, P* prev);
+    P* buildMatrixCell(int x,int y, int cost, P* prev);
 
  public:
     SearchableMatrix<P>* buildMatrix(list<string> data);
